Align reference implementation of "wait for all" with latest Web IDL spec
The spec of "get a promise for waiting for all" changed slightly when it was moved from the promise guide to Web IDL. Notably, the successSteps and failureSteps arguments were removed, since you can already achieve the same thing by reacting to the returned promise.

This updates the reference implementation to align with the definitions from the Web IDL specification. No normative changes.