Fix readable/writable file stream examples
These examples weren't closing the filehandle correctly (they called `close` on a nonexistent `fd` variable, rather than calling the `fileHandle.close` method).

Also the ReadableStream example misused the `fileHandle.read` call: it was supplying an ArrayBuffer and v's byteOffset within the ArrayBuffer, but the API requires supplying v itself and an offset relative to v's start.

Also added the file offset `position` to that `fileHandle.read` call. There's no reason otherwise for the example to keep track of this. Alternatively, the argument could be supplied as null, and we just rely on implicit file positioning.