Ensure pipeTo() releases its reader correctly
The user agent is allowed to use a default reader or a BYOB reader for the pipe, but we were always calling ReadableStreamDefaultReaderRelease in the finalize steps. This commit fixes the algorithm to we should use the appropriate release abstract operation corresponding to the chosen reader.

This was broken in d5f92d9f17306d31ba6b27424d23d58e89bf64a5.