Remove ReadableStreamBYOBReadResult in favour of a renamed ReadableStreamReadResult
In https://github.com/whatwg/streams/pull/1214, ReadableStreamBYOBReadResult was
changed to have a value property of type `(ArrayBufferView or undefined) value`,
since it is possible for the value to be undefined.

However, the WebIDL specification says that is invalid IDL:

> undefined must not be used as the type of an argument in any circumstance (in
> an operation, callback function, constructor operation, etc), or as the type
> of a dictionary member, whether directly or in a union. Instead, use an
> optional argument or a non-required dictionary member.
>
- https://webidl.spec.whatwg.org/#idl-undefined

So this commit instead removes ReadableStreamBYOBReadResult entirely, in favour
of a renamed ReadableStreamDefaultReadResult, which has type `any` for its value.