TransformStream cleanup using "Transformer.cancel"
This commit adds a "cancel" hook to "Transformer". This allows users to
perform resource cleanup when the readable side of the TransformStream
is cancelled, or the writable side is aborted.

To preserve existing behavior, when the readable side is cancelled with
a reason, the writable side is always immediately aborted with that same
reason. The same is true in the reverse case. This means that the
status of both sides is always either "closed", "erroring", or
"erroring" when the "cancel" hook is called.

"flush" and "cancel" are never both called. As per existing behaviour,
when the writable side is closed the "flush" hook is called. If the
readable side is cancelled while a promise returned from "flush" is
still pending, "cancel" is not called. In this scenario the readable
side ends up in the "errored" state, while the writable side ends up in
the "closed" state.